+++ foods/src/app/app.module.ts
@@ -1,5 +1,4 @@
-import { TabsComponent } from './components/tabs/tabs.component';
-import { LoginPage } from './pages/login/login.page';
+
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
@@ -17,11 +16,6 @@ import { HttpClientModule } from '@angular/common/http';
 import { PipesModule } from './pipes/pipes.module';
 import { Camera} from '@ionic-native/camera/ngx';
 
-import { AngularFireAuthModule  } from '@angular/fire/auth';
-import { AngularFireModule  } from '@angular/fire';
-import { GooglePlus } from '@ionic-native/google-plus/ngx';
-import { firebaseConfig } from 'src/environments/environment';
-
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
@@ -31,8 +25,6 @@ import { firebaseConfig } from 'src/environments/environment';
     AppRoutingModule,
     HttpClientModule,
     IonicStorageModule.forRoot(),
-    AngularFireAuthModule,
-    AngularFireModule.initializeApp( firebaseConfig )
   ],
   providers: [
     StatusBar,
@@ -40,7 +32,6 @@ import { firebaseConfig } from 'src/environments/environment';
     CategoriaService,
     PipesModule,
     Camera,
-    GooglePlus,
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
   ],
   bootstrap: [AppComponent]
