+++ foods/src/app/components/tabs/tabs.component.html
@@ -10,7 +10,7 @@
             <ion-label color="primary"> QR </ion-label>
         </ion-tab-button>
   
-        <ion-tab-button tab="/register">
+        <ion-tab-button tab="/foods/src/assets/sello1.jpg">
             <ion-icon name="person-outline" color="primary"></ion-icon>
             <ion-label color="primary"> Perfil </ion-label>
         </ion-tab-button>
