+++ foods/src/app/pages/login/login.page.html
@@ -1,6 +1,9 @@
+<ion-header>
+  <ion-toolbar color="warning">
+    <img src="/assets/UTPL-FONDO-AZUL.jpg">
+  </ion-toolbar>
+</ion-header>
 <ion-content>
-
-
   <ion-slides class="mainSlide" >
     
     <ion-slide >
@@ -10,10 +13,9 @@
 
           <ion-row>
             <ion-col>
-              <img src="/assets/avatars/av-1.png">
+              <img  src="/assets/avatars/av-1.png" style=" width: 120px;" >
             </ion-col>
           </ion-row>
-          
           <ion-row>
             <ion-col>
               
@@ -42,20 +44,25 @@
           <ion-row>
             <ion-col>
               <ion-button type="submit"
-                          color="tertiary"
-                          shape="round">
-                Login
+              shape="round"
+              expand="full"
+              color="primary">
+                Ingresar
               </ion-button>
             </ion-col>
+           
           </ion-row>
           <ion-row>
             <ion-col>
-              <ion-button type="submit"
-                          class="google"
-                          shape="round"
-                          (click)="loginGoogle()">
-                 <ion-icon name="logo-google" ></ion-icon>  Inicia Sesion con Google 
-              </ion-button>
+              <ion-col>
+                <ion-button type="submit"
+                shape="round"
+                expand="full"
+                color="primary"
+                            routerLink="/register">
+                  Registrarse
+                </ion-button>
+              </ion-col>
             </ion-col>
           </ion-row>
         </ion-grid>
@@ -65,114 +72,6 @@
 
     </ion-slide>
 
-    <ion-slide>
 
-      
-      <ion-grid fixed>
-        
-        <ion-row>
-          <ion-col>
-              <h3>Seleccione Avatar</h3>
-          </ion-col>
-          </ion-row>
-        
-          <ion-row>
-          <ion-col>
-              <ion-slides [options]="avatarSlide">
-              <ion-slide *ngFor="let avatar of avatars">
-                  <ion-img class="pick-avatar" 
-                          src="/assets/avatars/{{avatar.img}}" 
-                          [ngClass]="{'pick-avatar-seleccionado': avatar.seleccionado}" 
-                          (click)="SeleccionarAvatar( avatar )"
-                          (avatarSel)="registerUser.avatar = $event"></ion-img>
-              </ion-slide>
-              </ion-slides>
-          </ion-col>
-          </ion-row>
-
-            <form #fRegistro="ngForm" (ngSubmit)="registro(fRegistro)" >
-                <ion-row>
-                    <ion-col>
-                        
-                        <ion-list>
-                            
-                            <ion-item>
-                            <ion-label>Email</ion-label>
-                            <ion-input name="email" 
-                                        type="email"
-                                        [(ngModel)]="registerUser.email"
-                                        required></ion-input>
-                            </ion-item>
-
-                            <ion-item>
-                            <ion-label>Nombres</ion-label>
-                            <ion-input name="nombre"
-                                        type="text"
-                                        [(ngModel)]="registerUser.nombres"
-                                        required></ion-input>
-                            </ion-item>
-                            <ion-item>
-                              <ion-label>Usuario</ion-label>
-                              <ion-input name="usuario"
-                                          type="text"
-                                          [(ngModel)]="registerUser.usuario"
-                                          required></ion-input>
-                              </ion-item>
-  
-
-                            <ion-item>
-                            <ion-label>Password</ion-label>
-                            <ion-input name="password"
-                                        type="password"
-                                        [(ngModel)]="registerUser.password"
-                                        required></ion-input>
-                            </ion-item>
-
-                        </ion-list>
-                    </ion-col>
-                </ion-row>
-                
-                <ion-row>
-                <ion-col>
-                    <ion-button type="submit"
-                                color="tertiary"
-                                shape="round"
-                                >
-                        Crear usuario
-                    </ion-button>
-                </ion-col>
-            </ion-row>
-            </form>
-        </ion-grid>
-      </ion-slide>
   </ion-slides>
 </ion-content>
-
-
-<ion-footer no-border>
-  <ion-toolbar>
-    <ion-row>
-      <ion-col>
-          <ion-button shape="round"
-                      expand="full"
-                      size="small"
-                      fill="outline"
-                      color="tertiary"
-                      >
-            Ingresar
-          </ion-button>
-      </ion-col>
-
-      <ion-col>
-        <ion-button shape="round"
-                    expand="full"
-                    size="small"
-                    fill="outline"
-                    color="tertiary"
-                    >
-          Registrarme
-        </ion-button>
-      </ion-col>
-    </ion-row>
-  </ion-toolbar>
-</ion-footer>
