+++ foods/src/app/pages/register/register.page.html
@@ -1,0 +1,80 @@
+<ion-header>
+  <ion-toolbar color="warning">
+    <img src="/assets/UTPL-FONDO-AZUL.jpg">
+  </ion-toolbar>
+</ion-header>
+
+<ion-content>
+  <ion-slide>
+
+      
+    <ion-grid fixed>
+      
+      <ion-row>
+        <ion-col>
+            <h3>Registrarte</h3>
+        </ion-col>
+        </ion-row>
+      
+        <ion-row>
+        <ion-col>
+            
+        </ion-col>
+        </ion-row>
+
+          <form #fRegistro="ngForm" (ngSubmit)="registro(fRegistro)" >
+              <ion-row>
+                  <ion-col>
+                      
+                      <ion-list>
+                          
+                          <ion-item>
+                          <ion-label>Email</ion-label>
+                          <ion-input name="email" 
+                                      type="email"
+                                      [(ngModel)]="registerUser.email"
+                                      required></ion-input>
+                          </ion-item>
+
+                          <ion-item>
+                          <ion-label>Nombres</ion-label>
+                          <ion-input name="nombre"
+                                      type="text"
+                                      [(ngModel)]="registerUser.nombres"
+                                      required></ion-input>
+                          </ion-item>
+                          <ion-item>
+                            <ion-label>Usuario</ion-label>
+                            <ion-input name="usuario"
+                                        type="text"
+                                        [(ngModel)]="registerUser.usuario"
+                                        required></ion-input>
+                            </ion-item>
+
+
+                          <ion-item>
+                          <ion-label>Password</ion-label>
+                          <ion-input name="password"
+                                      type="password"
+                                      [(ngModel)]="registerUser.password"
+                                      required></ion-input>
+                          </ion-item>
+
+                      </ion-list>
+                  </ion-col>
+              </ion-row>
+              
+              <ion-row>
+              <ion-col>
+                  <ion-button type="submit"
+                              color="primary"
+                              shape="round"
+                              >
+                      Crear usuario
+                  </ion-button>
+              </ion-col>
+          </ion-row>
+          </form>
+      </ion-grid>
+    </ion-slide>
+</ion-content>
