+++ foods/src/app/pages/register/register.page.scss
@@ -3,9 +3,9 @@
     height: 100%;
 }
 
-.img {
-    width: 120px;
-}
+// img {
+//     width: 120px;
+// }
 
 .pick-avatar {
     width: 80px;
@@ -16,7 +16,6 @@
     transition: opacity 0.5s linear;
     opacity: 1 !important;
 }
-
 ion-header ion-toolbar {
     --background: url('../../../assets/logo.png'); 
 }
