+++ foods/src/app/pages/subcategorias/subcategorias.page.html
@@ -3,7 +3,7 @@
   <ion-buttons slot="start">
     <ion-back-button defaultHref="/"></ion-back-button>
   </ion-buttons>
-  <ion-title>Alimentos</ion-title>
+  <ion-title color="warning">Alimentos</ion-title>
 </ion-toolbar>
 <ion-content [fullscreen]="true" color="ion-color-medium-tint">
   <ion-grid fixed>
@@ -21,9 +21,3 @@
     </ion-row>
   </ion-grid>
 </ion-content>
-
-<ion-footer>
-  <ion-toolbar>
-    <app-tabs></app-tabs>
-  </ion-toolbar>
-</ion-footer>
