+++ foods/src/app/services/login.service.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Storage } from '@ionic/storage';
 import { IUsuario } from '../interface/usuario';
-import { AngularFireAuth } from '@angular/fire/auth';
 
 @Injectable({
   providedIn: 'root'
@@ -15,7 +14,7 @@ export class LoginService {
 
   constructor( private http: HttpClient,
                private storange: Storage,
-               private afAuth: AngularFireAuth) { }
+              ) { }
 
   login(usuario: string, password: string){
     const data = { usuario, password };
@@ -55,12 +54,12 @@ export class LoginService {
       });
   }
 
-  loginGoogle(usuario: string, password: string){
-    return new Promise( (resolve, rejects) => {
-          this.afAuth.signInWithEmailAndPassword( usuario, password).then(user => {
-          resolve(user);
-      }).catch(err =>
-          rejects(err));
-    });
-  }
+  // loginGoogle(usuario: string, password: string){
+  //   return new Promise( (resolve, rejects) => {
+  //         this.afAuth.signInWithEmailAndPassword( usuario, password).then(user => {
+  //         resolve(user);
+  //     }).catch(err =>
+  //         rejects(err));
+  //   });
+  // }
 }
